KPDF Shell: Instances can now be reused
This commit adds support for reusing existing KPDF instances to display externally opened documents in new tabs instead of new windows.
The feature depends on whether the corresponding option is checked and on whether KPDF is invoked with any URLs.

This commit adds a DCOP interface which contains tab-related methods.
Also breakoff tabs are now launched in a new instance instead of creating a new Shell widget in the same app, as this causes problems with the KPDFShellDCOPInterface. As a useful bonus, the new "--new-instance" CLI switch is now available to users and scripts.

Signed-off-by: Mavridis Philippe <mavridisf@gmail.com>
(cherry picked from commit 77ebecd92cef8d8c368ffb9cee3c856888456546)